merges visibleitemlist and itemlist, recreates functions as selectors
