+++ src/Kiosk/components/ItemList.jsx
@@ -1,5 +1,8 @@
 import React from 'react'
 import Item from './Item.jsx'
+import { connect } from 'react-redux'
+import { addItemToCart, openIngredientModalForItem } from '../actions'
+import { getSelectedCategory, getItems } from '../selectors'
 
 const ItemList = React.createClass({
     propTypes: {
@@ -39,4 +42,36 @@ const ItemList = React.createClass({
     }
 })
 
-export default ItemList
+const getVisibleItems = (items, category) => {
+    if (category === 0) {
+        return items
+    } else {
+        return items.filter((i) => i.category === category)
+    }
+}
+
+const mapStateToProps = (state) => {
+    return {
+        items: getVisibleItems(getItems(state), getSelectedCategory(state))
+    }
+}
+
+const mapDispatchToProps = (dispatch) => {
+    return {
+        // eslint-disable-next-line camelcase
+        onItemClick: (item, can_have_ingredients) => {
+            // eslint-disable-next-line camelcase
+            if (can_have_ingredients) {
+                $('#ingredient-modal').openModal()
+                dispatch(openIngredientModalForItem(item))
+            } else {
+                dispatch(addItemToCart(item))
+            }
+        }
+    }
+}
+
+export default connect(
+    mapStateToProps,
+    mapDispatchToProps
+)(ItemList)
